+++ internal/runner/parallel_runner.go
@@ -2,11 +2,11 @@ package runner
 
 import (
 	"context"
+	"github.com/alitto/pond"
 	"github.com/mih-kopylov/bulker/internal/config"
 	"github.com/mih-kopylov/bulker/internal/settings"
 	"github.com/sirupsen/logrus"
 	"github.com/spf13/afero"
-	"sync"
 )
 
 type ParallelRunner struct {
@@ -16,7 +16,7 @@ type ParallelRunner struct {
 	filter  *Filter
 }
 
-func (r ParallelRunner) Run(handler RepoHandler) error {
+func (r *ParallelRunner) Run(handler RepoHandler) error {
 	type repoProcessResult struct {
 		Name string
 		ProcessResult
@@ -30,34 +30,34 @@ func (r ParallelRunner) Run(handler RepoHandler) error {
 	ctx := context.Background()
 
 	allReposResult := map[string]ProcessResult{}
-	wg := sync.WaitGroup{}
+	pool := pond.New(r.config.MaxWorkers, 1000, pond.Context(ctx))
+	defer pool.StopAndWait()
 	ch := make(chan repoProcessResult)
 	logrus.WithField("mode", r.config.RunMode).Debug("processing repositories")
-	processedRepoCount := 0
-	for _, repo := range sets.Repos {
-		if !r.filter.Matches(repo) {
-			continue
-		}
-		processedRepoCount++
+	repos := r.filter.FilterMatchingRepos(sets.Repos)
+	progress := NewProgress(r.config, len(repos))
+	for _, repo := range repos {
 		runContext := newRunContext(r.fs, r.manager, r.config, repo)
-		wg.Add(1)
-		go func(ch chan repoProcessResult) {
-			defer wg.Done()
-			repoResult, err := handler(ctx, runContext)
-			ch <- repoProcessResult{
-				Name: runContext.Repo.Name,
-				ProcessResult: ProcessResult{
-					Result: repoResult,
-					Error:  err,
-				},
-			}
-		}(ch)
+		pool.Submit(
+			func() {
+				logrus.WithField("repo", runContext.Repo.Name).Debug("processing started")
+				repoResult, err := handler(ctx, runContext)
+				progress.Incr()
+				logrus.WithField("repo", runContext.Repo.Name).Debug("processing completed")
+				ch <- repoProcessResult{
+					Name: runContext.Repo.Name,
+					ProcessResult: ProcessResult{
+						Result: repoResult,
+						Error:  err,
+					},
+				}
+			},
+		)
 	}
-	for i := 0; i < processedRepoCount; i++ {
+	for i := 0; i < len(repos); i++ {
 		result := <-ch
 		allReposResult[result.Name] = result.ProcessResult
 	}
-	wg.Wait()
 	close(ch)
 
 	err = logOutput(allReposResult)